Protocols for Mappa, MD and ItemP
