+++ skytemple_files/common/ppmdu_config/dungeon_data.py
@@ -73,7 +73,7 @@ class Pmd2DungeonItemCategory(AutoString):
 
     def item_ids(self, only_if_valid_in_mappa=True):
         if only_if_valid_in_mappa:
-            from skytemple_files.dungeon_data.mappa_bin.item_list import MAX_ITEM_ID
+            from skytemple_files.dungeon_data.mappa_bin._python_impl.item_list import MAX_ITEM_ID
 
             return [x for x in self.items if x < MAX_ITEM_ID]
         return self.items
