+++ skytemple_files/data/item_p/test/__init__.py
@@ -1,4 +1,3 @@
-"""Converts ItemP models back into the binary format used by the game"""
 #  Copyright 2020-2022 Capypara and the SkyTemple Contributors
 #
 #  This file is part of SkyTemple.
@@ -15,23 +14,3 @@
 #
 #  You should have received a copy of the GNU General Public License
 #  along with SkyTemple.  If not, see <https://www.gnu.org/licenses/>.
-from __future__ import annotations
-
-from typing import Optional
-
-from skytemple_files.common.util import *
-from skytemple_files.data.item_p import *
-from skytemple_files.data.item_p.model import ItemP
-
-
-class ItemPWriter:
-    def __init__(self, model: ItemP):
-        self.model = model
-
-    def write(self) -> Tuple[bytes, List[int], Optional[int]]:
-        pointer_offsets: List[int] = []
-        header_offset = 0
-        data = bytearray(0)
-        for i in self.model.item_list:
-            data += i.to_bytes()
-        return bytes(data), pointer_offsets, header_offset
