+++ skytemple_files/data/md/test/__init__.py
@@ -14,6 +14,3 @@
 #
 #  You should have received a copy of the GNU General Public License
 #  along with SkyTemple.  If not, see <https://www.gnu.org/licenses/>.
-
-
-from __future__ import annotations
