+++ skytemple_files/dungeon_data/mappa_bin/_python_impl/floor.py
@@ -22,7 +22,6 @@ from xml.etree.ElementTree import Element
 from skytemple_files.common.i18n_util import _, f
 from skytemple_files.common.util import *
 from skytemple_files.common.xml_util import (
-    XmlSerializable,
     XmlValidateError,
     validate_xml_tag,
 )
@@ -41,13 +40,14 @@ from skytemple_files.dungeon_data.mappa_bin import (
     XML_MONSTER_LIST,
     XML_TRAP_LIST,
 )
-from skytemple_files.dungeon_data.mappa_bin.floor_layout import MappaFloorLayout
-from skytemple_files.dungeon_data.mappa_bin.item_list import MappaItemList
-from skytemple_files.dungeon_data.mappa_bin.monster import MappaMonster
-from skytemple_files.dungeon_data.mappa_bin.trap_list import MappaTrapList
+from skytemple_files.dungeon_data.mappa_bin._python_impl.floor_layout import MappaFloorLayout
+from skytemple_files.dungeon_data.mappa_bin._python_impl.item_list import MappaItemList
+from skytemple_files.dungeon_data.mappa_bin._python_impl.monster import MappaMonster
+from skytemple_files.dungeon_data.mappa_bin._python_impl.trap_list import MappaTrapList
+from skytemple_files.dungeon_data.mappa_bin.protocol import MappaFloorProtocol
 
 if TYPE_CHECKING:
-    from skytemple_files.dungeon_data.mappa_bin.model import MappaBinReadContainer
+    from skytemple_files.dungeon_data.mappa_bin._python_impl.model import MappaBinReadContainer
 
 
 class StubMappaFloor:
@@ -107,7 +107,7 @@ class StubMappaFloor:
         return data
 
 
-class MappaFloor(AutoString, XmlSerializable):
+class MappaFloor(MappaFloorProtocol[MappaFloorLayout, MappaMonster, MappaTrapList, MappaItemList], AutoString):
     def __init__(
         self,
         layout: MappaFloorLayout,
@@ -222,147 +222,6 @@ class MappaFloor(AutoString, XmlSerializable):
     def _from_cache(read, pnt, load_callback):
         # TODO: Caching needs a deep copy
         return load_callback(pnt)
-
-    def to_xml(
-        self,
-        export_layout=True,
-        export_monsters=True,
-        export_traps=True,
-        export_floor_items=True,
-        export_shop_items=True,
-        export_monster_house_items=True,
-        export_buried_items=True,
-        export_unk1_items=True,
-        export_unk2_items=True,
-    ) -> Element:
-        floor_xml = Element(XML_FLOOR)
-
-        if export_layout:
-            layout_xml = self.layout.to_xml()
-            validate_xml_tag(layout_xml, XML_FLOOR_LAYOUT)
-            floor_xml.append(layout_xml)
-
-        if export_monsters:
-            monsters_xml = Element(XML_MONSTER_LIST)
-            for monster in self.monsters:
-                monster_xml = monster.to_xml()
-                validate_xml_tag(monster_xml, XML_MONSTER)
-                monsters_xml.append(monster_xml)
-            floor_xml.append(monsters_xml)
-
-        if export_traps:
-            traps_xml = self.traps.to_xml()
-            validate_xml_tag(traps_xml, XML_TRAP_LIST)
-            floor_xml.append(traps_xml)
-
-        if export_floor_items:
-            floor_items_xml = self.floor_items.to_xml()
-            validate_xml_tag(floor_items_xml, XML_ITEM_LIST)
-            floor_items_xml.set(XML_ITEM_LIST__TYPE, XML_ITEM_LIST__TYPE__FLOOR)
-            floor_xml.append(floor_items_xml)
-
-        if export_shop_items:
-            shop_items_xml = self.shop_items.to_xml()
-            validate_xml_tag(shop_items_xml, XML_ITEM_LIST)
-            shop_items_xml.set(XML_ITEM_LIST__TYPE, XML_ITEM_LIST__TYPE__SHOP)
-            floor_xml.append(shop_items_xml)
-
-        if export_monster_house_items:
-            monster_house_items_xml = self.monster_house_items.to_xml()
-            validate_xml_tag(monster_house_items_xml, XML_ITEM_LIST)
-            monster_house_items_xml.set(
-                XML_ITEM_LIST__TYPE, XML_ITEM_LIST__TYPE__MONSTER_HOUSE
-            )
-            floor_xml.append(monster_house_items_xml)
-
-        if export_buried_items:
-            buried_items_xml = self.buried_items.to_xml()
-            validate_xml_tag(buried_items_xml, XML_ITEM_LIST)
-            buried_items_xml.set(XML_ITEM_LIST__TYPE, XML_ITEM_LIST__TYPE__BURIED)
-            floor_xml.append(buried_items_xml)
-
-        if export_unk1_items:
-            unk_items1_xml = self.unk_items1.to_xml()
-            validate_xml_tag(unk_items1_xml, XML_ITEM_LIST)
-            unk_items1_xml.set(XML_ITEM_LIST__TYPE, XML_ITEM_LIST__TYPE__UNK1)
-            floor_xml.append(unk_items1_xml)
-
-        if export_unk2_items:
-            unk_items2_xml = self.unk_items2.to_xml()
-            validate_xml_tag(unk_items2_xml, XML_ITEM_LIST)
-            unk_items2_xml.set(XML_ITEM_LIST__TYPE, XML_ITEM_LIST__TYPE__UNK2)
-            floor_xml.append(unk_items2_xml)
-
-        return floor_xml
-
-    @classmethod
-    @typing.no_type_check
-    def from_xml(cls, ele: Element) -> "MappaFloor":
-        data = {
-            "layout": None,
-            "monsters": None,
-            "traps": None,
-            "floor_items": None,
-            "shop_items": None,
-            "monster_house_items": None,
-            "buried_items": None,
-            "unk_items1": None,
-            "unk_items2": None,
-        }
-        for child in ele:
-            if child.tag == XML_FLOOR_LAYOUT and data["layout"] is None:
-                data["layout"] = MappaFloorLayout.from_xml(child)
-            elif child.tag == XML_MONSTER_LIST and data["monsters"] is None:
-                monsters = []
-                for monster in child:
-                    monsters.append(MappaMonster.from_xml(monster))
-                data["monsters"] = monsters
-            elif child.tag == XML_TRAP_LIST and data["traps"] is None:
-                data["traps"] = MappaTrapList.from_xml(child)
-            elif (
-                child.tag == XML_ITEM_LIST
-                and child.get(XML_ITEM_LIST__TYPE) == XML_ITEM_LIST__TYPE__FLOOR
-                and data["floor_items"] is None
-            ):
-                data["floor_items"] = MappaItemList.from_xml(child)
-            elif (
-                child.tag == XML_ITEM_LIST
-                and child.get(XML_ITEM_LIST__TYPE) == XML_ITEM_LIST__TYPE__SHOP
-                and data["shop_items"] is None
-            ):
-                data["shop_items"] = MappaItemList.from_xml(child)
-            elif (
-                child.tag == XML_ITEM_LIST
-                and child.get(XML_ITEM_LIST__TYPE) == XML_ITEM_LIST__TYPE__MONSTER_HOUSE
-                and data["monster_house_items"] is None
-            ):
-                data["monster_house_items"] = MappaItemList.from_xml(child)
-            elif (
-                child.tag == XML_ITEM_LIST
-                and child.get(XML_ITEM_LIST__TYPE) == XML_ITEM_LIST__TYPE__BURIED
-                and data["buried_items"] is None
-            ):
-                data["buried_items"] = MappaItemList.from_xml(child)
-            elif (
-                child.tag == XML_ITEM_LIST
-                and child.get(XML_ITEM_LIST__TYPE) == XML_ITEM_LIST__TYPE__UNK1
-                and data["unk_items1"] is None
-            ):
-                data["unk_items1"] = MappaItemList.from_xml(child)
-            elif (
-                child.tag == XML_ITEM_LIST
-                and child.get(XML_ITEM_LIST__TYPE) == XML_ITEM_LIST__TYPE__UNK2
-                and data["unk_items2"] is None
-            ):
-                data["unk_items2"] = MappaItemList.from_xml(child)
-            else:
-                raise XmlValidateError(f(_("Floor parsing: Unexpected {child.tag}")))
-
-        for k, v in data.items():
-            if v is None:
-                raise XmlValidateError(f(_("Missing {k} for Floor data.")))
-
-        return cls(**data)
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, MappaFloor):
