+++ skytemple_files/dungeon_data/mappa_bin/_python_impl/item_list.py
@@ -26,7 +26,6 @@ from skytemple_files.common.ppmdu_config.dungeon_data import (
 )
 from skytemple_files.common.util import *
 from skytemple_files.common.xml_util import (
-    XmlSerializable,
     XmlValidateError,
     validate_xml_attribs,
     validate_xml_tag,
@@ -40,28 +39,21 @@ from skytemple_files.dungeon_data.mappa_bin import (
     XML_ITEM__WEIGHT,
     XML_ITEM_LIST,
 )
-from skytemple_files.dungeon_data.mappa_bin._deprecated import MappaItemCategory
+from skytemple_files.dungeon_data.mappa_bin.protocol import MappaItemListProtocol, Probability, _MappaItemCategory, \
+    _MappaItem, MAX_ITEM_ID, CMD_SKIP, GUARANTEED
 
 if TYPE_CHECKING:
-    from skytemple_files.dungeon_data.mappa_bin.model import MappaBinReadContainer
+    from skytemple_files.dungeon_data.mappa_bin._python_impl.model import MappaBinReadContainer
 logger = logging.getLogger(__name__)
 
 
-CMD_SKIP = 0x7530
-GUARANTEED = 0xFFFF
-MAX_ITEM_ID = 363
-POKE_ID = 183
-# Actually GUARANTEED or a weight between 0 and MAX_WEIGHT.
-Probability = int
-
-
-class MappaItemList(AutoString, XmlSerializable):
+class MappaItemList(MappaItemListProtocol, AutoString):
     def __init__(
         self,
         categories: Dict[
-            Union[MappaItemCategory, Pmd2DungeonItemCategory], Probability
+            _MappaItemCategory, Probability
         ],
-        items: Dict[Pmd2DungeonItem, Probability],
+        items: Dict[_MappaItem, Probability],
     ):
         self.categories = categories
         self.items = items
@@ -71,7 +63,7 @@ class MappaItemList(AutoString, XmlSerializable):
         return cls.from_bytes(read.data, read.items, pointer)
 
     @classmethod
-    def from_bytes(cls, data: bytes, item_list: List[Pmd2DungeonItem], pointer: int):
+    def from_bytes(cls, data: bytes, item_list: List[_MappaItem], pointer: int) -> "MappaItemList":
         processing_categories = True
         item_or_cat_id = 0
         orig_pointer = pointer
@@ -107,78 +99,32 @@ class MappaItemList(AutoString, XmlSerializable):
 
         return MappaItemList(categories, items)  # type: ignore
 
-    def to_mappa(self):
+    def to_mappa(self) -> bytes:
         data = bytearray()
         current_id = 0
         # Start with the categories
-        for cat, val in sorted(self.categories.items(), key=lambda it: it[0].value):
-            id_cat = cat.value
+        for cat, val in sorted(self.categories.items(), key=lambda it: it[0]):
+            id_cat = cat
             if current_id != id_cat:
                 current_id = self._write_skip(data, current_id, id_cat)
             self._write_probability(data, val)
             current_id += 1
         # Continue with the items
-        sorted_items = sorted(self.items.items(), key=lambda it: it[0].id)
-        first_item_id = sorted_items[0][0].id if len(sorted_items) > 0 else 0
+        sorted_items = sorted(self.items.items(), key=lambda it: it[0])
+        first_item_id = sorted_items[0][0] if len(sorted_items) > 0 else 0
         self._write_skip(data, current_id, 0x10 + first_item_id)
         current_id = first_item_id
         for item, val in sorted_items:
-            if current_id != item.id:
-                current_id = self._write_skip(data, current_id, item.id)
+            if current_id != item:
+                current_id = self._write_skip(data, current_id, item)
             self._write_probability(data, val)
             current_id += 1
         # Fill up to MAX_ITEM_ID + 1
         self._write_skip(data, current_id, MAX_ITEM_ID + 1)
         return data
 
-    def to_xml(self) -> Element:
-        xml_item_list = Element(XML_ITEM_LIST)
-        for category, probability in self.categories.items():
-            weight = "GUARANTEED" if probability == GUARANTEED else str(probability)
-            xml_category = Element(
-                XML_CATEGORY,
-                {
-                    XML_CATEGORY__NAME: MappaItemCategory(category.value).name,  # type: ignore
-                    XML_CATEGORY__WEIGHT: str(weight),
-                },
-            )
-            xml_item_list.append(xml_category)
-        for item, probability in self.items.items():
-            weight = "GUARANTEED" if probability == GUARANTEED else str(probability)
-            xml_item = Element(
-                XML_ITEM, {XML_ITEM__ID: str(item.id), XML_ITEM__WEIGHT: str(weight)}
-            )
-            xml_item_list.append(xml_item)
-
-        return xml_item_list
-
-    @classmethod
-    @typing.no_type_check
-    def from_xml(cls, ele: Element) -> "XmlSerializable":
-        validate_xml_tag(ele, XML_ITEM_LIST)
-        categories = {}
-        items = {}
-        for child in ele:
-            if child.tag == XML_CATEGORY:
-                validate_xml_attribs(child, [XML_CATEGORY__NAME, XML_CATEGORY__WEIGHT])
-                name = child.get(XML_CATEGORY__NAME)
-                # TODO: Switch to Pmd2DungeonItemCategory
-                if not hasattr(MappaItemCategory, name):
-                    raise XmlValidateError(f"Unknown item category {name}.")
-                weight_str = child.get(XML_CATEGORY__WEIGHT)
-                weight = int(weight_str) if weight_str != "GUARANTEED" else GUARANTEED
-                # TODO: Switch to Pmd2DungeonItemCategory
-                categories[getattr(MappaItemCategory, name)] = weight
-            elif child.tag == XML_ITEM:
-                validate_xml_attribs(child, [XML_ITEM__ID, XML_ITEM__WEIGHT])
-                weight_str = child.get(XML_ITEM__WEIGHT)
-                weight = int(weight_str) if weight_str != "GUARANTEED" else GUARANTEED
-                items[Pmd2DungeonItem(int(child.get(XML_ITEM__ID)), "???")] = weight
-            else:
-                raise XmlValidateError(
-                    f"Unexpected sub-node for {XML_ITEM_LIST}: {child.tag}"
-                )
-        return cls(categories, items)
+    def to_bytes(self) -> bytes:
+        return self.to_mappa()
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, MappaItemList):
