+++ skytemple_files/hardcoded/dungeons.py
@@ -22,7 +22,7 @@ import typing
 from skytemple_files.common.i18n_util import _
 from skytemple_files.common.ppmdu_config.data import Pmd2Data
 from skytemple_files.common.util import *
-from skytemple_files.data.md.model import PokeType
+from skytemple_files.data.md.protocol import PokeType
 
 DUNGEON_LIST_ENTRY_LEN = 4
 DUNGEON_RESTRICTIONS_ENTRY_LEN = 12
